py-requirements.txt
dexml
image
imageio
matplotlib
numpy
pillow
scipy
scikit-learn
scikit-image
